Split logic out of view
